hurricane-proc-send-data/requirements.txt
pandas==1.2.2
requests==2.25.1
python-dotenv==0.15.0
fire==0.4